Add cloudinary external image storage

+++ blog/views.py
@@ -1,10 +1,12 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
 from django.views import generic
 from .models import Blog
 from . import forms, models
 from django.views.generic import ListView 
 from django.contrib.auth.decorators import login_required
 from django.forms import formset_factory
+from django.contrib.auth.models import User
 
 
 # Create your views here.
@@ -120,24 +122,27 @@ def edit_blog(request, blog_id):
     blog = get_object_or_404(models.Blog, id=blog_id)
     edit_form = forms.BlogForm(instance=blog)
     delete_form = forms.DeleteBlogForm()
-    if request.method == 'POST':
-        if 'edit_blog' in request.POST:
-            edit_form = forms.BlogForm(request.POST, instance=blog)
-            if edit_form.is_valid():
-                edit_form.save()
-                return redirect('home')
-        
-        if 'delete_blog' in request.POST:
-            delete_form = forms.DeleteBlogForm(request.POST)
-            if delete_form.is_valid():
-                blog.delete()
-                return redirect('home')
-    context = {
-        'edit_form': edit_form,
-        'delete_form': delete_form,
-    }
-    return render(request, 'blog/edit_blog.html', context=context)
-
+    
+    if request.user == blog.author:
+        if request.method == 'POST':
+            
+            if 'edit_blog' in request.POST:
+                edit_form = forms.BlogForm(request.POST, instance=blog)
+                if edit_form.is_valid():
+                    edit_form.save()
+                    return redirect('home')
+            
+            if 'delete_blog' in request.POST:
+                delete_form = forms.DeleteBlogForm(request.POST)
+                if delete_form.is_valid():
+                    blog.delete()
+                    return redirect('home')
+        context = {
+            'edit_form': edit_form,
+            'delete_form': delete_form,
+        }
+        return render(request, 'blog/edit_blog.html', context=context)
+    return HttpResponse('The logged-in user is not the author of this blog post.')
 
  
 class PostList(ListView):
